// src/app/enums/game-state.enum.ts
export enum GameState {
  // ALWAYS START THE LOOP
  FIGHT_INIT = 'FIGHT_INIT',
  // LOOP IN THESES STATE STILL ENEDED
  TURN_DECIDE = 'TURN_DECIDE',
  PLAYER_TURN = 'PLAYER_TURN',
  ENEMY_TURN = 'ENEMY_TURN',
  APPLY_EFFECT = 'APPLY_EFFECT',
  CHECK_END = 'CHECK_END',
  // ALWAYS END THE FIGHT -> LOOP ENDED
  FIGHT_END = 'FIGHT_END',
  // WHEN NOT IN FIGHT PHASE
  NONE = 'NONE'
}